Moved all calculators to a designated folder
