+++ applications/GeoMechanicsApplication/geo_mechanics_application.h
@@ -89,7 +89,7 @@
 #include "custom_elements/U_Pw_small_strain_interface_element.h"
 #include "custom_elements/U_Pw_updated_lagrangian_FIC_element.h"
 #include "custom_elements/U_Pw_updated_lagrangian_element.h"
-#include "custom_elements/calculation_contribution.h"
+#include "custom_elements/contribution_calculators/calculation_contribution.h"
 #include "custom_elements/geo_steady_state_Pw_piping_element.h"
 #include "custom_elements/small_strain_U_Pw_diff_order_element.h"
 #include "custom_elements/steady_state_Pw_element.h"
